Crate logger

A default logger for Libra project.

Usage

use logger::prelude::*;

pub fn main() {
    let _g = logger::set_default_global_logger(false /* async */, Some(256));
    info!("Starting...");
}

Modules

prelude
Logger prelude which includes all logging macros.

Macros

crit
Log a critical level message using current scope logger

debug
Log a debug level message using current scope logger

error
Log a error level message using current scope logger

info
Log a info level message using current scope logger

log_collector_crit
Log a critical level message using current log collector

log_collector_debug
Log a debug level message using current log collector

log_collector_error
Log a error level message using current log collector

log_collector_info
Log a info level message using current log collector

log_collector_trace
Log a trace level message using current log collector

log_collector_warn
Log a warning level message using current log collector

set_default_global_logger
Set up a local binding to the default logger.

slog_crit
Log critical level record (alias)

slog_debug
Log debug level record (alias)

slog_error
Log error level record

slog_info
Log info level record (alias)

slog_trace
Log trace level record (alias)

slog_warn
Log warning level record (alias)

trace
Log a trace level message using current scope logger

warn
Log a warning level message using current scope logger

Structs

GLOBAL_LOG_COLLECTOR

Functions

create_default_root_logger
Creates a root logger with default settings.

create_env_logger
Creates a logger that respects RUST_LOG environment variable

create_env_logger_with_level
Creates a logger that respects RUST_LOG environment variable

create_test_root_logger
Creates a root logger with test settings: does not do output if test passes. Caveat: cargo test does not capture output for non main thread. So this logger is not very useful for multithreading scenarios.

get_log_collector

init_for_e2e_testing
Create and setup default global logger for use in end to end testing.

set_default_global_logger
Creates and sets default global logger. Caller must keep the returned guard alive.

set_global_log_collector

try_init_for_testing
Create and setup default global logger following the env-logger conventions, i.e. configured by environment variable RUST_LOG. This is useful to make logging optional in unit tests.

with_logger
Access the Global Logger Collector for the current logging scope