Function logger::init

pub fn init<I, W>(
    log_level: Option<Level>, 
    writers: impl IntoIterator<Item = I>
) -> Result<(), SetLoggerError> where
    W: Write + Send + 'static,
    I: Into<Arc<MutexIrqSafe<W>>>, 

Initialize the fully-featured Theseus system logger.

Arguments

log_level: the log level that should be used. If None, the DEFAULT_LOG_LEVEL will be used.

writers: an iterator over the backends that the system logger will write log messages to. Typically this is just a single writer, such as the COM1 serial port.