Function logger::set_log_level

pub fn set_log_level(level: Level)

Set the log level, which determines whether a given log message is actually logged.

For example, if Level::Trace is set, all log levels will be logged.

If Level::Info is set, debug!() and trace!() will not be logged, but info!(), warn!(), and error!() will be.