Crate logger

Custom logger module.

A custom logger implementation over log crate that uses stdout to print logs.

To make clear, log crate is a facade for the logging facilities in Rust. And user ue its macros (such as log::debug! or log::info!) to log messages. But log itself will not print out or write any logs. Instead, it will use a specified logger implementation, such as Logger, to do so. It is the actual implementation that defines format, style and destination of the logs.

So for binary targets, both the log crate and the Logger implementation should be added as dependencies.

# Cargo.toml for binary targets
[dependencies]
log = {version = "0.4", features = ["std"]}
logger = { path = "/path/to/logger" }

However, for library targets, only the log crate should be added as a dependency. This is because the log will not be generated until it’s used in a binary target. And the actual logger implementation used in binary will be used in this scenario.

So for a library target, this logger implementation is actually not a dependency.

# Cargo.toml for library targets
[dependencies]
log = {version = "0.4", features = ["std"]}

And similarly, doing logging in the library is as simple as using the log crate.

log::debug!("This is a debug message in a library");

Note that log crate should be set with "sdt" feature. This apply both to library and binary targets.

Structs

Logger

A custom logger struct that uses stdout to print logs.