Crate logger

logger API suitable for C

While we eventually want to move the implementation to Rust, Rust code should generally use Rust’s log crate instead.

Since we use this in the shim, we eventually want this to not depend on std nor libc. (The C part of the implementation currently uses libc). https://github.com/shadow/shadow/issues/2919

Structs

_Logger

__va_list_tag

Constants

_LogLevel_LOGLEVEL_DEBUG

_LogLevel_LOGLEVEL_ERROR

_LogLevel_LOGLEVEL_INFO

_LogLevel_LOGLEVEL_TRACE

_LogLevel_LOGLEVEL_UNSET

_LogLevel_LOGLEVEL_WARNING

Functions

logger_abort⚠

logger_base_name⚠

logger_elapsed_micros⚠

logger_elapsed_string⚠

logger_flush⚠

logger_getDefault⚠

logger_get_global_start_time_micros⚠

logger_isEnabled⚠

logger_log⚠

logger_now_micros⚠

logger_setDefault⚠

logger_setLevel⚠

logger_set_global_start_time_micros⚠

Type Aliases

LogLevel

Logger

_LogLevel

__builtin_va_list

va_list