Crate logger

A basic logger implementation for system-wide logging in Theseus.

This enables Theseus crates to use the log crate’s macros anywhere, such as error!(), warn!(), info!(), debug!(), and trace!().

Currently, log statements are written to one or more writers, which are objects that implement the core::fmt::Write trait.

Constants

DEFAULT_LOG_LEVEL
By default, Theseus will print all log levels, including Trace and above.

LOG_MAX_WRITERS
The maximum number of output streams that a logger can write to.

Functions

early_init
Initializes Theseus’s early system logger, which only supports logging to basic serial ports.

init
Initialize the fully-featured Theseus system logger.

mirror_to_vga
Call this to enable mirroring logging macros to the screen

set_log_level
Set the log level, which determines whether a given log message is actually logged.

take_early_log_writers
Removes all of the writers (output streams) from the early logger and returns them.

write_fmt
Convenience function for writing formatted arguments to the logger.

write_str
Convenience function for writing a simple string to the logger.

Type Definitions

LogOutputFunc
The signature of a callback function that will optionally be invoked on every log statement to be printed, which enables log mirroring. See mirror_to_vga().