Crate logger

A basic logger implementation for system-wide logging in Theseus.

This enables Theseus crates to use the log crate’s macros anywhere, such as error!(), warn!(), info!(), debug!(), and trace!().

Currently, log statements are written to one or more writers, which are objects that implement the core::fmt::Write trait.

Early log messages (before memory management is initialized) are saved to a static fixed-sized buffer such that they are not lost and can be retrieved once logging sinks are ready to be used.

Constants

DEFAULT_LOG_LEVEL

By default, Theseus will print all log levels, including Trace and above.

EARLY_LOG_BUFFER_SIZE

The size of the buffer used to save early log messages.

LOG_MAX_WRITERS

The maximum number of output streams that a logger can write to.

Functions

early_init

Initializes Theseus’s early system logger for use before memory management is set up.

init

Initialize the fully-featured Theseus system logger.

set_log_level

Set the log level, which determines whether a given log message is actually logged.

take_early_log_writers

Removes all of the writers (output streams) from the early logger and returns them.

write_fmt

Convenience function for writing formatted arguments to the logger.

write_str

Convenience function for writing a simple string to the logger.