Function logger::early_init

pub fn early_init(
    log_level: Option<Level>,
    serial_ports: impl IntoIterator<Item = SerialPort>
) -> Result<(), SetLoggerError>

Initializes Theseus’s early system logger, which only supports logging to basic serial ports.

Arguments

log_level: the log level that should be used. If None, the DEFAULT_LOG_LEVEL will be used.

serial_ports: an iterator of [SerialPort]s that the logger will write log messages to. Typically this is just a single serial port, e.g., COM1.

This function will initialize the logger with a maximum of LOG_MAX_WRITERS serial ports; any additional ones in the given iterator beyond that will be ignored.