Struct logger::Logger

pub struct Logger { /* private fields */ }

A custom logger struct that uses stdout to print logs.

To use it in a binary target, first initialize the logger with a logging level. Then use the log macros to log messages.

extern crate log;
use crate::logger::Logger;
use log::LevelFilter::Debug;

fn main() {
    Logger::new().set_level(Debug).init();
    log::debug!("This is a debug message");
}

Also, use the set_prefix() method to enable a prefix for the logger.

Implementations

impl Logger

pub fn new() -> Logger

Create a new global logger with default level set to LevelFilter::Trace.

use crate::logger::Logger;
Logger::new().init();
log::info!("Logger initialized");

Remember to call init() method to initialize the logger after creating and configuring it.

pub fn init(self)

Static method to initialize the logger with an optional logging level.

If no level is set via set_level, LevelFilter::Trace will be used as default level.

use crate::logger::Logger;
Logger::new().init();
log::info!("Logger initialized");

pub fn set_level(self, level: LevelFilter) -> Logger

Set the default level of the logger.

use crate::logger::Logger;
use log::LevelFilter::Warn;
Logger::new().set_level(Warn).init();
log::info!("Logger initialized.");

Remember to call init() method to initialize the logger after creating and configuring it.

pub fn set_prefix(self, prefix: String) -> Logger

Enable the logger with a prefix after the time.

use crate::logger::Logger;
use log::LevelFilter::Warn;
let prefix = String::from("test_logger");
Logger::new().set_prefix(prefix).init();
log::info!("Logger initialized.");

Remember to call init() method to initialize the logger after creating and configuring it.

pub fn get_level(&self) -> LevelFilter

Get the current level of the logger.

pub fn get_prefix(&self) -> String

Get the current prefix of the logger. If no prefix is set, it will return an empty string.

pub fn is_using_prefix(&self) -> bool

Get the current status of the prefix. If no prefix is set, it will return false. Otherwise, it will return true.